requirements.txt
boto3==1.13.7
celery==4.4.0
celery-redbeat==0.11.1
cryptography==2.9.2
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==3.0.7
django-anymail[sendinblue]==7.1.0
django-cache-memoize==0.1.6
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
django-fontawesome-5
django-measurement==3.2.3
django-money==1.1
django-redis-cache==2.1.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-request==1.5.5
django-simple-history==2.10.0
django-storages==1.9.1
django-timezone-field
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
drf-jwt==1.16.1
iommi==0.7.0
ipython==7.14.0
line-bot-sdk
measurement==3.2.0
# psycopg2
psycopg2-binary
python-dotenv==0.13.0
python-jose==3.1.0
py-trello==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.1
requests-oauthlib==1.3.0
requests==2.23.0
schedule==0.6.0
simplejson==3.17.0
social-auth-app-django==3.1.0
six==1.14.0
toolz==0.10.0
waitress==1.4.3

#filters for django-markup
bleach==3.1.4
bleach-whitelist==0.0.10
html5lib==1.0.1
markdown==3.1
python-creole==1.3.2
smartypants==2.0.1
textile==3.0.3
webencodings==0.5.1
